requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=3.0.0
ldap3>=2.9
python-dotenv>=0.19.0
APScheduler>=3.9.0
Flask-WTF>=0.15.1
zeep>=4.2.1
python-docx>=0.8.11
PyMuPDF>=1.19.0
docx2pdf>=0.1.8
